refactor: adjusts board component file structure
