+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Board from './components/Board/Board';
+import Board from './components/Board/Board/Board';
 
 export default function App() {
   return <Board />;
