+++ src/components/Board/Rank/Rank.types.ts
@@ -1,5 +1,9 @@
-import { BoardRank, BoardFullRank, BoardSquare } from '../../utils/board.types';
-import { BoardData, BoardHandlers } from './Board.types';
+import {
+  BoardRank,
+  BoardFullRank,
+  BoardSquare,
+} from '../../../utils/board.types';
+import { BoardData, BoardHandlers } from '../Board/Board.types';
 
 export interface RankProps {
   number: BoardRank;
